Implemented dynamic routing and optimized page structure
