+++ package.json
@@ -30,6 +30,8 @@
   },
   "packageManager": "yarn@3.2.3",
   "devDependencies": {
+    "@types/react-modal": "^3.16.0",
+    "@types/react-transition-group": "^4",
     "dotenv": "^16.0.3",
     "husky": "^8.0.1",
     "lint-staged": "^13.0.3"
@@ -37,6 +39,9 @@
   "dependencies": {
     "@uniswap/v3-periphery": "^1.4.3",
     "chart.js": "^4.3.0",
-    "ethers": "^5.7.2"
+    "ethers": "^5.7.2",
+    "react-modal": "^3.16.1",
+    "react-toastify": "^9.1.3",
+    "react-transition-group": "^4.4.5"
   }
 }
