--- packages/nextjs/components/subscription/SelectSubscription.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-export function SelectSubscription() {
-  return (
-    <div>
-      <h1 className="mb-4 font-semibold text-gray-900 dark:text-white">Create a subscription</h1>
-      <h1 className="mb-4 font-semibold text-gray-900 dark:text-white">Selection an asset </h1>
-    </div>
-  );
-}
