+++ packages/nextjs/components/subscription/SubscriptionType.tsx
@@ -1,90 +1,66 @@
-import React from "react";
+import React, { useState } from "react";
+import Image from "next/image";
+
+interface Subscription {
+  id: number;
+  title: string;
+  imageSrc: string;
+}
 
 export function SubscriptionType() {
+  const [selectedSubscription, setSelectedSubscription] = useState<number | null>(null);
+
+  const handleSubscriptionChange = (subscriptionId: number) => {
+    setSelectedSubscription(subscriptionId);
+    console.log(subscriptionId);
+  };
+
+  const subscriptions: Subscription[] = [
+    {
+      id: 1,
+      title: "Time based",
+      imageSrc: "/assets/img/3d-hourglass.svg",
+    },
+    {
+      id: 2,
+      title: "Inflation rate",
+      imageSrc: "/assets/img/inflation.svg",
+    },
+    {
+      id: 3,
+      title: "Asset price",
+      imageSrc: "/assets/img/3d-sold.svg",
+    },
+    {
+      id: 4,
+      title: "Trending volume",
+      imageSrc: "/assets/img/trending.svg",
+    },
+  ];
+
   return (
     <div>
-      <h1 className="mb-4 font-semibold text-gray-900 dark:text-white">Create a subscription</h1>
-      <h1 className="mb-4 font-semibold text-gray-900 dark:text-white">Please select subscription type </h1>
+      <h1 className="text-sm font-medium mb-2">Create a subscription</h1>
+      <h2 className="text-2xl font-bold mb-12">Please select subscription type</h2>
+
+      <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-2 lg:grid-cols-4 gap-6">
+        {subscriptions.map(subscription => (
+          <div className="border border-gray-300 p-4 relative" key={subscription.id}>
+            <label className="absolute top-2 right-2">
+              <input
+                type="radio"
+                name="subscription"
+                value={subscription.id}
+                checked={selectedSubscription === subscription.id}
+                onChange={() => handleSubscriptionChange(subscription.id)}
+                className="form-radio h-4 w-4 text-blue-500 cursor-pointer"
+              />
+            </label>
+            <Image src={subscription.imageSrc} alt={subscription.title} width={300} height={200} className="mb-2" />
+            <p className=" font-medium text-center">{subscription.title}</p>
+          </div>
+        ))}
+      </div>
     </div>
   );
 }
-
-// import React from "react";
-
-// export function SubscriptionCheckBox() {
-//   return (
-//     <div>
-//       <h3 className="mb-4 font-semibold text-gray-900 dark:text-white">Identification</h3>
-//       <ul className="items-center w-full text-sm font-medium text-gray-900 bg-white border border-gray-200 rounded-lg sm:flex dark:bg-gray-700 dark:border-gray-600 dark:text-white">
-//         <li className="w-full border-b border-gray-200 sm:border-b-0 sm:border-r dark:border-gray-600">
-//           <div className="flex items-center pl-3">
-//             <input
-//               id="horizontal-list-radio-license"
-//               type="radio"
-//               value=""
-//               name="list-radio"
-//               className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-700 dark:focus:ring-offset-gray-700 focus:ring-2 dark:bg-gray-600 dark:border-gray-500"
-//             />
-//             <label
-//               htmlFor="horizontal-list-radio-license"
-//               className="w-full py-3 ml-2 text-sm font-medium text-gray-900 dark:text-gray-300"
-//             >
-//               Driver License{" "}
-//             </label>
-//           </div>
-//         </li>
-//         <li className="w-full border-b border-gray-200 sm:border-b-0 sm:border-r dark:border-gray-600">
-//           <div className="flex items-center pl-3">
-//             <input
-//               id="horizontal-list-radio-id"
-//               type="radio"
-//               value=""
-//               name="list-radio"
-//               className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-700 dark:focus:ring-offset-gray-700 focus:ring-2 dark:bg-gray-600 dark:border-gray-500"
-//             />
-//             <label
-//               htmlFor="horizontal-list-radio-id"
-//               className="w-full py-3 ml-2 text-sm font-medium text-gray-900 dark:text-gray-300"
-//             >
-//               State ID
-//             </label>
-//           </div>
-//         </li>
-//         <li className="w-full border-b border-gray-200 sm:border-b-0 sm:border-r dark:border-gray-600">
-//           <div className="flex items-center pl-3">
-//             <input
-//               id="horizontal-list-radio-millitary"
-//               type="radio"
-//               value=""
-//               name="list-radio"
-//               className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-700 dark:focus:ring-offset-gray-700 focus:ring-2 dark:bg-gray-600 dark:border-gray-500"
-//             />
-//             <label
-//               htmlFor="horizontal-list-radio-millitary"
-//               className="w-full py-3 ml-2 text-sm font-medium text-gray-900 dark:text-gray-300"
-//             >
-//               US Millitary
-//             </label>
-//           </div>
-//         </li>
-//         <li className="w-full dark:border-gray-600">
-//           <div className="flex items-center pl-3">
-//             <input
-//               id="horizontal-list-radio-passport"
-//               type="radio"
-//               value=""
-//               name="list-radio"
-//               className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-700 dark:focus:ring-offset-gray-700 focus:ring-2 dark:bg-gray-600 dark:border-gray-500"
-//             />
-//             <label
-//               htmlFor="horizontal-list-radio-passport"
-//               className="w-full py-3 ml-2 text-sm font-medium text-gray-900 dark:text-gray-300"
-//             >
-//               US Passport
-//             </label>
-//           </div>
-//         </li>
-//       </ul>
-//     </div>
-//   );
-// }
