+++ packages/nextjs/components/subscription/index.tsx
@@ -1,3 +1,2 @@
 export * from "./ActiveSubscriptions";
-export * from "./SelectSubscription";
 export * from "./SubscriptionItem";
